add Permission checking at DeleteItem Component and backend

+++ backend/src/resolvers/Mutation.js
@@ -51,7 +51,14 @@ const Mutations = {
     // 1. find the item
     const item = await ctx.db.query.item({ where }, `{ id title}`);
     // 2. Check if they own that item, or have the permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
